rename date-picker.js -> date-label.js
